% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_uni.R
\name{approx_uni}
\alias{approx_uni}
\title{Linearly interpolate non-unique numbers}
\usage{
approx_uni(x, y, xint)
}
\arguments{
\item{x}{numeric of x locations bounding the interpolation}

\item{y}{numeric of y locations bounding the interpolation}

\item{xint}{numeric x location to interpolate y}
}
\value{
The \code{y} value if all values are non-unique, otherwise an interpolated value defined by \code{xint}
}
\description{
Linearly interpolate non-unique numbers, a modification to the standard \code{\link[stats]{approx}} function.
}
\details{
Used in \code{\link{resinterp}}
}

