% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chlinterp.R
\name{chlinterp}
\alias{chlinterp}
\alias{chlinterp.numeric}
\title{Interpolate from a fit grid}
\usage{
chlinterp(row_in, ...)

\method{chlinterp}{numeric}(row_in, sal_pred, sal_grd, ...)
}
\arguments{
\item{row_in}{vector of observations from a row of an interpolation grid}

\item{...}{arguments passed to additional methods}

\item{sal_pred}{value of salff for interpolating chlorophyll}

\item{sal_grd}{original vector of salff values used to create interpolation grid}
}
\value{
Predicted chlorophyll values from the interpolation
}
\description{
Interpolate from a fit grid for a given salinity value (as fraction of freshwater) and observation (row)
}
\details{
This function is primarily used within \code{\link{chlprd}} and \code{\link{chlnorm}}.
}
\examples{
##

# load data
data(tidobj)

# get a row from the interpolation grid to predict
fit_grd <- attr(tidobj, 'fits')
fit_grd <- fit_grd[['fit0.5']]
fit_row <- fit_grd[1, ]

# get salinity values used to create interpolation grid
sal_grd <- attr(tidobj, 'sal_grd')

# interpolate chlorophyll from an observed salinity value
}

