% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chllab.R
\name{chllab}
\alias{chllab}
\title{Chlorophyll axis label}
\usage{
chllab(logspace = TRUE)
}
\arguments{
\item{logspace}{logical indicating if chlorophyll is in log space, otherwise observed}
}
\value{
An expression indicating (log)-chlorophyll in the appropriate units
}
\description{
Get the chlorophyll axis label for observed or log-space, including units
}
\examples{
## default
chllab()
}

