% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chlnormmean.R
\name{chlnormmean}
\alias{chlnormmean}
\alias{chlnormmean.tidal}
\title{Get salinity normalized WRTDS predictions from interpolation grids for the mean model}
\usage{
chlnormmean(tidal_in, ...)

\method{chlnormmean}{tidal}(tidal_in, trace = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to or from other methods}

\item{trace}{logical indicating if progress is shown in the console}
}
\value{
Appends columns to the data.frame of normalized chlorophyll values for the mean model in log-space and the observed values from the back-transformed grids.  Columns are named as `nrms_mean' and `btnrms_mean'.
}
\description{
Get normalized model predictions from WRTDS for the mean model to remove the effect of salinity on chlorophyll.  Predicted values in the interpolation grids are averaged across dates.
}
\details{
This function is used after \code{wrtdsmean} to normalize predicted and back-transformed predicted values of chlorophyll from the interpolation grids for the mean model.  The normalized values are based on the average of all predicted chlorophyll estimates across the range of salinity values that have occurred on the same date throughout each year.  For example, normalized values for July 2000 are the mean predicted chlorophyll at that date using the observed salinity values that occur in July of all years.  The normalized values allow an interpretation of chlorophyll trends that are independent of changes in salinity or freshwater inputs.
}
\examples{
##

# load a tidal object
data(tidobjmean)

# get flow-normalized values
res <- chlnormmean(tidobjmean)

##
\dontrun{
# fails if incorrect input
data(chldat)
chldat <- tidal(chldat)
chlnormmean(chldat)
}
}

