% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chlpred.R
\name{chlpred}
\alias{chlpred}
\alias{chlpred.tidal}
\title{Get WRTDS predictions from interpolation grids}
\usage{
chlpred(tidal_in, ...)

\method{chlpred}{tidal}(tidal_in, ...)
}
\arguments{
\item{tidal_in}{input tidal object}
}
\value{
Appends columns to the data.frame of predicted chlorophyll value for each quantile.
}
\description{
Get model predictions from WRTDS using linear interpolation of values in grids
}
\details{
This function is used after \code{wrtds} to estimate predicted values of chlorophyll from the interpolation grid.  The estimated values are based on a linear interpolation of the two predicted chlorophyll values in the interpolation grid that are bounded by the upper and lower bounds of salinity for the observed salinity at each observation.
}
\examples{
##

# load a tidal object
data(tidobj)

chlpred(tidobj)
}

