% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chlpredmean.R
\name{chlpredmean}
\alias{chlpredmean}
\alias{chlpredmean.tidal}
\title{Get WRTDS predictions for the mean model from interpolation grids}
\usage{
chlpredmean(tidal_in, ...)

\method{chlpredmean}{tidal}(tidal_in, trace = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to or from other methods}

\item{trace}{logical indicating if progress is shown in the console}
}
\value{
Appends columns to the data.frame of predicted chlorophyll values for the mean model in log-space and the observed values from the back-transformed grids.  Columns are named as `fits_mean' and `btfits_mean'.
}
\description{
Get model predictions from WRTDS for the mean model using linear interpolation of values in grids
}
\details{
This function is used after \code{wrtdsmean} to estimate predicted values of chlorophyll from the interpolation grid of the mean model.  The estimated values are based on a linear interpolation of the two predicted chlorophyll values in the interpolation grid that are bounded by the upper and lower bounds of salinity for the observed salinity at each observation.
}
\examples{
##

# load a tidal object
data(tidobjmean)

# get predicted values
res <- chlpredmean(tidobjmean)

##
\dontrun{
# fails if incorrect input
data(chldat)
chldat <- tidal(chldat)
chlpredmean(chldat)
}
}

