% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chngest.R
\name{chngest}
\alias{chngest}
\title{Get trend for a single time period}
\usage{
chngest(x, y, single = F)
}
\arguments{
\item{x}{numeric of year values}

\item{y}{norm numeric of annually averaged response variable}

\item{single}{logical if trends are based on only first and last year of aggregated, i.e., no averaging of first/last three}
}
\description{
Get trend for a single time period
}
\details{
Estimates trends using methods described in \code{\link{wrtdstrnd}}.  Used internally and is not to be called by the user. Function runs on individual time period groups defined by arguments in \code{\link{wrtdstrnd}}.
}

