% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillpo.R
\name{fillpo}
\alias{fillpo}
\alias{fillpo.tidal}
\alias{fillpo.tidalmean}
\title{Fill the \code{predonobs} attribute}
\usage{
fillpo(dat_in, ...)

\method{fillpo}{tidal}(dat_in, ...)

\method{fillpo}{tidalmean}(dat_in, ...)
}
\arguments{
\item{dat_in}{input tidal or tidalmean object}

\item{...}{arguments passed to or from other methods}
}
\value{
The input tidal or tidalmean object with the filled \code{predonobs} attribute as predictions for the observed data as a data frame.
}
\description{
Fill the \code{predonobs} attribute
}
\details{
Used in \code{\link{respred}} to fill the \code{predonobs} attribute.  This attribute is used to estimate performance metrics with \code{\link{wrtdsperf}}.
}
