% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{fitplot}
\alias{fitplot}
\alias{fitplot.tidal}
\title{Plot a tidal object}
\usage{
fitplot(tidal_in, ...)

\method{fitplot}{tidal}(tidal_in, tau = NULL, annuals = TRUE,
  logspace = FALSE, pretty = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to other methods}

\item{annuals}{logical indicating if plots are annual aggregations of results}

\item{logspace}{logical indicating if plots are in log space}

\item{pretty}{logical indicating if my subjective idea of plot aesthetic is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot a tidal object to view chlorophyll observations, predictions, and normalized results.
}
\examples{
## load a fitted tidal object
data(tidfit)

# plot
fitplot(tidfit)
}

