% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flofind.R
\name{flofind}
\alias{flofind}
\alias{flofind.default}
\title{Get all salinity or flow values across years for a common month}
\usage{
flofind(dat_in, ...)

\method{flofind}{default}(dat_in, obs, ...)
}
\arguments{
\item{dat_in}{input tidal object}

\item{...}{arguments passed to additional methods}

\item{obs}{index of row in tidal object to identify salinity or flow values}
}
\value{
Numeric vector of all observed salinity or flow values for a single month across all years
}
\description{
Get all salinity or flow values across years for a common month for normalizing predictions of the response variable
}
\details{
This function is used in \code{\link{resnorm}}.  It will return all observed salinity/flow values for a given date for matching months across all years.  For example, the function will return all observed salinity/flow values for January across all years in the dataset if the input row is for January 2000.  An identical vector would result for the same month in other years.  Salinity/flow values are averaged when more than one observation was taken in the same month for a given year.
}
\examples{
##

# load data
data(tidobj)

# find common salinity/flow values across years for the first observation
res <- flofind(tidobj, 1)
}

