% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getwts.R
\name{getwts}
\alias{getwts}
\alias{getwts.tidal}
\title{Get weights for regression}
\usage{
getwts(tidal_in, ...)

\method{getwts}{tidal}(tidal_in, ref_in, wt_vars = c("day_num", "year",
  "salff"), wins = list(0.5, 10, NULL), all = F, min_obs = T, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to other methods}

\item{ref_in}{row of tidal object as reference for weights}

\item{wt_vars}{chr string of three elements indicatings names of columns in tidal object that are used for reference row weights}

\item{wins}{list of half-window widths for time, year, and salinity}

\item{all}{logical to return individual weights rather than the product of all three, default \code{FALSE}}

\item{min_obs}{logical to use window widening if less than 100 non-zero weights are found, default \code{TRUE}}
}
\value{
Some weights
}
\description{
Get weights for weighted regression for a single observation using a tri-cubic weighting function
}
\examples{
##
}

