% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gridplot.R
\name{gridplot}
\alias{gridplot}
\alias{gridplot.tidal}
\title{Plot chlorophyll response to salinity as a gridded surface}
\usage{
gridplot(tidal_in, ...)

\method{gridplot}{tidal}(tidal_in, month = 7, tau = NULL, years = NULL,
  logspace = FALSE, allsal = FALSE, interp = TRUE, sal_fac = 5,
  yr_fac = sal_fac, pretty = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to other methods}

\item{month}{numeric input from 1 to 12 indicating the monthly predictions to plot}

\item{tau}{numeric vector of quantile to plot.  The function will plot the 'middle' quantile if none is specified, e.g., if 0.2, 0.3, and 0.4 are present in the fitted model object then 0.3 will be plotted.}

\item{years}{numeric vector of years to plot, defaults to all}

\item{logspace}{logical indicating if plots are in log space}

\item{allsal}{logical indicating if the salinity values for plotting are limited to the fifth and ninety-fifth percentile of observed salinity values for the month of interest}

\item{interp}{logical indicating if chlorophyll between integer years and salinity is linearly interpolated to create a smoother grid}

\item{sal_fac}{numeric value indicating the factor for smoothing chlorophyll across salinity values. Increasing the value creates more smoothing and setting the value to 1 removes all smoothing.}

\item{yr_fac}{numeric value indicating the factor for smoothing chlorophyll across integer years. Increasing the value creates more smoothing and setting the value to 1 removes all smoothing.}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot the relationship between chlorophyll and salinity across the time series using a gridded surface.  This can be used to evaluate temporal variation between the two.
}
\details{
The plot can be used to examine how the relationship between chlorophyll and salinity varies throughout the time series.  It is essentially identical to the plot produced by \code{\link{dynaplot}}, except a gridded plot is returned that shows salinity over time with cells colored by chlorophyll.  The interpolation grid that is stored as an attribute in a fitted tidal object is used to create the plot.  The plot is limited to the same month throughout the time series to limit seasonal variation.  By default, the plot is constrained to the fifth and ninety-fifth percentile of observed salinity values during the month of interest to limit the predictions within the data domain. This behavior can be suppressed by changing the \code{allsal} argument, although the predicted chlorophyll values that are outside of the salinity range for the plotted month are typically unrealistic.
}
\examples{
## load a fitted tidal object
data(tidfit)

## plot using defaults,
## defaults to the fiftieth quantile for July for all years
gridplot(tidfit)

## change the defaults
gridplot(tidfit, tau = 0.9, month = 2, years = seq(1980, 1990), sal_fac = 20)
}
\seealso{
\code{\link{dynaplot}}, \code{\link{fitplot}}, \code{\link{prdnrmplot}}
}

