% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modfitmean.R
\name{modfitmean}
\alias{modfitmean}
\alias{modfitmean.data.frame}
\title{Fit weighted regression and get predicted/normalized chlorophyll for the conditional mean model}
\usage{
modfitmean(dat_in, ...)

\method{modfitmean}{data.frame}(dat_in, ...)
}
\arguments{
\item{dat_in}{input \code{\link[base]{data.frame}} for fitting the model, see details}

\item{...}{arguments passed to or from other methods}
}
\value{
A tidal object with predicted and normalized chlorophyll predictions in log-space and back-transformed.
}
\description{
Fit weighted regression and get predicted/normalized chlorophyll for the conditioanl mean model from a data frame.  This is a wrapper for multiple function used to create a weighted regression model.
}
\details{
This function is used as a convenience to combine several functions that accomplish specific tasks, primarily the creation of a tidal object, fitting of the weighted regression models with \code{\link{wrtdsmean}}, extraction of fitted values from the interpolation grids using \code{\link{chlpredmean}}, and normalization of the fitted values from the interpolation grids using \code{\link{chlnormmean}}.  The format of the input data can be a \code{\link[base]{data.frame}} or \code{\link{tidal}} object, where the latter inherits methods from the former.  The data format should be chlorophyll observations as rows and the first four columns as date, chlorophyll, salinity, and detection limits.  The order of the columns may vary provided the order of each of the variables is specified by the \code{ind} argument that is passed to the \code{\link{tidal}} function.  The chlorophyll data are also assumed to be in log-space, otherwise use \code{chllog = FALSE} which is also passed to the \code{\link{tidal}} function.  Numerous other arguments affect the output and the default parameters may not be appropriate for most scenarios.  Arguments used by other functions can be specified explicitly with the initial call.  The documentation for the functions under `see also' should be consulted for available arguments, as well as the examples that illustrate common changes to the default values.
}
\examples{
\dontrun{
## load data
data(chldat)

## fit the model and get predicted/normalized chlorophyll data
# default median fit
# grids predicted across salinity range with ten values
res <- modfitmean(chldat)

## fit different quantiles and smaller interpolation grid
res <- modfitmean(chldat, tau = c(0.2, 0.8), sal_div = 5)

## fit with different window widths
# half-window widths of one day, five years, and 0.3 salinity
res <- modfitmean(chldat, wins = list(1, 5, 0.3))

## suppress console output
res <- modfitmean(chldat, trace = FALSE)
}
}
\seealso{
See the help files for \code{\link{tidal}}, \code{\link{wrtdsmean}}, \code{\link{getwts}}, \code{\link{chlpredmean}}, and \code{\link{chlnormmean}} for arguments that can be passed to this function.
}

