% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tidal}
\alias{plot.tidal}
\title{Plot a tidal object}
\usage{
\method{plot}{tidal}(tidal_in, annuals = TRUE, logspace = FALSE)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{annuals}{logical indicating if plots are annual aggregations of results}

\item{logspace}{logical indicating if plots are in log space}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot a tidal object to view chlorophyll observations, predictions, and normalized results.
}
\examples{
## load a fitted tidal object
data(tidfit)

# plot
plot(tidfit)
}

