% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prdnrmplot.R
\name{prdnrmplot}
\alias{prdnrmplot}
\alias{prdnrmplot.tidal}
\title{Plot combined predicted and normalized results from a tidal object}
\usage{
prdnrmplot(tidal_in, ...)

\method{prdnrmplot}{tidal}(tidal_in, tau = NULL, annuals = FALSE,
  logspace = FALSE, dt_rng = NULL, lwd = 1, size = 2, alpha = 1,
  pretty = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to \code{\link[ggplot2]{geom_line}}}

\item{tau}{numeric vector of quantiles to plot, defaults to all in object if not supplied}

\item{annuals}{logical indicating if plots are annual aggregations of results}

\item{logspace}{logical indicating if plots are in log space}

\item{dt_rng}{Optional chr string indicating the date range of the plot. Must be two values in the format 'YYYY-mm-dd' which is passed to \code{\link{as.Date}}.}

\item{lwd}{numeric value indicating width of lines}

\item{size}{numeric value indicating size of points}

\item{alpha}{numeric value indicating transparency of points or lines}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot combined predicted and normalized results from a tidal object to evaluate the influence of salinity changes on chlorophyll. The plot is similar to that produced by \code{\link{fitplot}} except predicted chlorophyll values are shown as points and observed values are removed.
}
\examples{
## load a fitted tidal object
data(tidfit)

# plot using defaults
prdnrmplot(tidfit)

# get the same plot but use default ggplot settings
prdnrmplot(tidfit, pretty = FALSE)

# plot in log space
prdnrmplot(tidfit, logspace = TRUE)

# plot specific quantiles
prdnrmplot(tidfit, tau = c(0.1, 0.9))

# plot the normalized predictions
prdnrmplot(tidfit, predicted = FALSE)

# plot as annual aggregations
prdnrmplot(tidfit, annuals = TRUE)

# format the x-axis is using annual aggregations
library(ggplot2)

prdnrmplot(tidfit, annual = TRUE) +
 scale_x_continuous(breaks = seq(2000, 2012, by = 4))

# modify the plot as needed using ggplot scales, etc.
prdnrmplot(tidfit, pretty = FALSE, linetype = 'dashed') +
 theme_classic() +
 scale_y_continuous(
   'Chlorophyll',
   limits = c(0, 50)
   ) +
 scale_colour_manual(
   '',
   labels = c('lo', 'md', 'hi'),
   values = c('red', 'green', 'blue'),
   guide = guide_legend(reverse = TRUE)
   )
}
\seealso{
\code{\link{fitplot}}, \code{\link{sliceplot}}
}

