% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resinterp.R
\name{resinterp}
\alias{resinterp}
\alias{resinterp.default}
\title{Interpolate from a fit grid}
\usage{
resinterp(date_in, flo_in, ...)

\method{resinterp}{default}(date_in, flo_in, fit_grd, flo_grd, ...)
}
\arguments{
\item{date_in}{obseved data to interpolate for the response variable}

\item{flo_in}{observed salinity/flow value to interpolate for the response variable}

\item{...}{arguments passed to additional methods}

\item{fit_grd}{interpolation grid created from \code{\link{wrtds}}}

\item{flo_grd}{original vector of salinity/flow values used to create interpolation grid}
}
\value{
Predicted values of the response variable from the interpolation
}
\description{
Interpolate the response variable in two dimensions from the fit grid, where the first dimension is salinity/flow and the second is date.
}
\details{
This function is used in \code{\link{respred}} and \code{\link{resnorm}}.
}
\examples{
##

# load data
data(tidobj)

# interpolation grid from the model
fit_grd <- attr(tidobj, 'fits')
fit_grd <- fit_grd[['fit0.5']]

# salinity/flow values used for interpolation grid
flo_grd <- attr(tidobj, 'flo_grd')

# get observed salinity/flow value to interpolate the response
date_in <- tidobj$date[1]
flo_in <- tidobj$flo[1]

# interpolate the response from an observed date and salinity/flow value
resinterp(date_in, flo_in, fit_grd, flo_grd)
}

