% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_sim.R
\name{samp_sim}
\alias{samp_sim}
\title{Sample a daily time series at a set frequency}
\usage{
samp_sim(dat_in, month_samps = 1)
}
\arguments{
\item{dat_in}{input \code{\link[base]{data.frame}} that is returned from \code{\link{lnres_sim}} or \code{\link{all_sims}}}

\item{month_samps}{numeric indicating sampling interval, defaults to once a month and is random within the month.  Multiple samples per month are based on sampling with replacement.}
}
\value{
Original data frame with rows subset based on number of desired monthly samples.
}
\description{
Sample a daily water quality time series at a set monthly frequency
}
\details{
This function is intended for sampling a simulated daily time series of water quality that is returned by \code{\link{lnres_sim}} or \code{\link{all_sims}}.
}
\examples{
\dontrun{
## example data
data(daydat)

## simulate
tosamp <- all_sims(daydat)

## sample
samp_sim(tosamp)
}
}
\seealso{
\code{\link{lnres_sim}}, \code{\link{all_sims}}
}

