% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sliceplot.R
\name{sliceplot}
\alias{sliceplot}
\alias{sliceplot.tidal}
\title{Plot time slices within a tidal object}
\usage{
sliceplot(tidal_in, ...)

\method{sliceplot}{tidal}(tidal_in, slices = c(1, 7), tau = NULL,
  predicted = TRUE, logspace = FALSE, pretty = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to \code{\link[ggplot2]{geom_line}}}

\item{slices}{numeric vector of calender months to plot, i.e., 1 - 12}

\item{tau}{numeric vector of quantile to plot, defaults to median quantile if none specified}

\item{predicted}{logical indicating if standard predicted values are plotted, default \code{TRUE}, otherwise normalized predictions are plotted}

\item{logspace}{logical indicating if plots are in log space}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot time slices within a tidal object to view chlorophyll observations, predictions, and normalized results at regular annual intervals.
}
\details{
Only one quantile fit can be shown per plot because the grouping variable is mapped to the slices.
}
\examples{
## load a fitted tidal object
data(tidfit)

# plot using defaults
sliceplot(tidfit)

# change x-axis labelling
library(ggplot2)
sliceplot(tidfit) +
 scale_x_continuous(breaks = seq(2000, 2012, by = 4))

# get different months - march and september
sliceplot(tidfit, slices = c(3, 9))

# normalized chlorophyll predictions, 10th percentile
sliceplot(tidfit, tau = 0.1, predicted = FALSE)

# normalized values all months, change line aesthetics, log-space, 90th
# add title
sliceplot(tidfit,
 slices = 1:12,
 size = 1.5,
 tau = 0.9,
 alpha = 0.6,
 predicted = FALSE,
 logspace = TRUE
) +
ggtitle('Normalized predictions for all months, 90th percentile')
}

