% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/winsrch_optim.R
\name{winsrch_optim}
\alias{winsrch_optim}
\alias{winsrch_optim.default}
\title{Quickly find the optimal half-window width combination}
\usage{
winsrch_optim(dat_in, ...)

\method{winsrch_optim}{default}(dat_in, wins_in = NULL, reltol = 0.001,
  lower = c(0.1, 1, 0.1), upper = c(2, 15, 2), ...)
}
\arguments{
\item{dat_in}{input data object to use with weighted regression}

\item{...}{arguments passed to or from other methods}

\item{wins_in}{starting list of window weights for initializing the search algorithm}

\item{reltol}{numeric for relative convergence tolerance passed to \code{control} argument of \code{\link[stats]{optim}}}

\item{lower}{vector of minimum half-window widths to evaluate}

\item{upper}{vector of maximum half-window widths to evaluate}
}
\value{
Some stuff
}
\description{
Quickly find the optimal half-window width combination to use for weighted regression
}
\details{
This function uses \code{\link[stats]{optim}} to minimize the error returned by \code{\link{wrtdscv}} for a given window combination.  The search algorithm uses the limited-memory modification of the BFGS quasi-Newton method to impose upper and lower limits on the optimization search. These limits can be changed using the \code{lower} and \code{upper} arguments.
}
\examples{
\dontrun{
# setup parallel backend
library(doParallel)
ncores <- detectCores() - 1
registerDoParallel(cores = ncores)

# run search function - takes a while
res <- winsrch_optim(tidobjmean)
}
}
\seealso{
\code{\link{wrtdscv}}, \code{\link{winsrch_grid}}
}

