% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrtdsmean.R
\name{wrtdsmean}
\alias{wrtdsmean}
\alias{wrtdsmean.data.frame}
\alias{wrtdsmean.tidal}
\title{Get WRTDS prediction grid for the mean response}
\usage{
wrtdsmean(tidal_in, sal_div = 10, trace = TRUE, ...)

\method{wrtdsmean}{data.frame}(tidal_in, sal_div = 10, trace = TRUE, ...)

\method{wrtdsmean}{tidal}(tidal_in, sal_div = 10, trace = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{sal_div}{numeric indicating number of divisions across the range of salinity to create the interpolation grid}

\item{trace}{logical indicating if progress is shown in the console}

\item{...}{arguments passed to or from other methods}
}
\value{
Appends a regression grid as an attribute to a tidal object
}
\description{
Get WRTDS prediction grid for the mean response of chlorophyll observations in a tidal object. The approach is similar to the \code{\link{wrtds}} function except that the mean response is modelled rather than specified quantiles.
}
\examples{
\dontrun{
## load data
data(chldat)

## get wrtds grid, as data frame input
res <- wrtdsmean(chldat)

## as tidal object
tidal_in <- tidal(chldat)
res <- wrtdsmean(tidal_in)

}
}

