% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrtdsres.R
\name{wrtdsres}
\alias{wrtdsres}
\alias{wrtdsres.tidal}
\title{Get WRTDS residuals}
\usage{
wrtdsres(tidal_in, ...)

\method{wrtdsres}{tidal}(tidal_in, trace = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object which must already have fitted model data}

\item{...}{arguments passed to or from other methods}

\item{trace}{logical indicating if progress is shown in the console}
}
\value{
Columns are added to the data of the tidal object for the residuals and non-conditional residuals of each observation.  This process is not as simple as subtracting observed minus predicted values for two reasons.  First, the residual at each observation is specific to the weighted regression at the center of the window.  This means that the WRTDS models must be refit to get the residuals.  Second, the goodness of fit measure described in Koenker and Machado (1999) for quantile regression requires a non-conditional quantile for comparison with the conditional quantile.  The non-conditional quantile is not as simple as estimating the relevant quantile for all data in the response since the weights are not considered as during the model fitting process.  Overall, the function is not as computationally intense as \code{\link{wrtds}} since values are only estimated for the predictions as compared to the entire interpolation grid for a quantile.

A tidal object with columns for the residuals ('res') and non-conditional residuals ('resnl') of each quantile model.
}
\description{
Get WRTDS residuals for each quantile model. These are used to estimate goodness of fit of the model predictions.
}
\examples{
\dontrun{
## load a fitted model object
data(tidfit)

## run the function
wrtdsres(tidfit)
}
}
\references{
Koenker, R., Machado, J.A.F. 1999. Goodness of fit and related inference processes for quantile regression. Journal of the American Statistical Association. 94(448):1296-1310.
}
\seealso{
\code{\link{wrtds}}
}

