% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollfun.R
\name{wrtdsroll}
\alias{wrtdsroll}
\title{Function passed to \code{\link[zoo]{rollapply}} in \code{\link{wrtds2}}}
\usage{
wrtdsroll(x, dat_in)
}
\arguments{
\item{x}{row}

\item{dat_in}{input data frame that is rolled}
}
\description{
Function passed to \code{\link[zoo]{rollapply}} in \code{\link{wrtds2}}
}

