% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrtdstrnd.R
\name{wrtdstrnd}
\alias{wrtdstrnd}
\alias{wrtdstrnd.default}
\alias{wrtdstrnd.tidal}
\alias{wrtdstrnd.tidalmean}
\title{Get WRTDS trends}
\usage{
wrtdstrnd(dat_in, ...)

\method{wrtdstrnd}{default}(dat_in, mobrks, yrbrks, molabs, yrlabs,
  aves = FALSE, ...)

\method{wrtdstrnd}{tidal}(dat_in, mobrks, yrbrks, molabs, yrlabs, tau = NULL,
  aves = FALSE, ...)

\method{wrtdstrnd}{tidalmean}(dat_in, mobrks, yrbrks, molabs, yrlabs,
  aves = FALSE, ...)
}
\arguments{
\item{dat_in}{input tidal or tidalmean object which must already have fitted model data}

\item{...}{methods passed to or from other methods}

\item{mobrks}{numeric vector of breaks for months, see examples}

\item{yrbrks}{numeric vector of breaks for years, see examples}

\item{molabs}{character vector of names for month breaks, see examples}

\item{yrlabs}{character vector of names for year breaks, see examples}

\item{aves}{logical if averages within each period are also returned}

\item{tau}{numeric vector of quantile for estimating trends}
}
\value{
A \code{\link[base]{data.frame}} with summary trends for each grouping
}
\description{
Get WRTDS trends for annual and monthly groupings
}
\details{
Trends are reported as percent changes of annual averages from the beginning to the end of each period.  To reduce the effects of odd years at the beginning and end of each period, percent changes are based on an average of the first three and last three annual averages.  For example, percent changes for January throughout an an entire time series from 1980 to 2000 would be the change of the average from January in 1980-1982 to the average from January in 1998-2000.  Annual trends, e.g., percent changes from 1980-1986, 1987-1993, etc. do not average by the first and last three years in each grouping because the values are already based on annual averages.

Averages in each period can be returned if \code{aves = TRUE}.  These averages are based on annual averages within each period for congruency with the trend estimates.  

All trends are based on back-transformed, flow-normalized results. 

The user must supply the annual and monthly aggregation periods to the appropriate arguments. These are passed to \code{\link[base]{cut}} and are left-open, right-closed along the interval.
}
\examples{
## load a fitted model object
data(tidfit)
data(tidfitmean)

## get trends

# setup month, year categories
mobrks <- c(-Inf, 3, 6, 9, Inf)
yrbrks <- c(-Inf, 1985, 1994, 2003, Inf)
molabs <- c('JFM', 'AMJ', 'JAS', 'OND')
yrlabs <- c('1974-1985', '1986-1994', '1995-2003', '2004-2012')

wrtdstrnd(tidfit, mobrks, yrbrks, molabs, yrlabs)
wrtdstrnd(tidfitmean, mobrks, yrbrks, molabs, yrlabs)

# get averages in each period
wrtdstrnd(tidfit, mobrks, yrbrks, molabs, yrlabs, aves = TRUE)
}

