% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wtsplot.R
\name{wtsplot}
\alias{wtsplot}
\alias{wtsplot.tidal}
\title{Plot the weights for an observation}
\usage{
wtsplot(tidal_in, ...)

\method{wtsplot}{tidal}(tidal_in, ref, wins = list(0.5, 10, NULL),
  logspace = FALSE, pretty = TRUE, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to other methods}

\item{ref}{chr string indicating the date at the center of the weighting window, must be in the format 'YYYY-mm-dd' which is passed to \code{\link{as.Date}}}

\item{wins}{list with three elements passed to \code{\link{getwts}} indicating the half-window widhts for day, year, and salinity}

\item{logspace}{logical indicating if plots are in log space}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Create several plots showing the weights used to fit a model for a single observation.
}
\details{
Create diagnostic plots to view the effects of different weighting windows on model predictions.
}
\examples{
## load a fitted tidal object
data(tidfit)

## plot using defaults,
wtsplot(tidfit, ref = '2000-07-01')

## change the defaults
}
\seealso{
\code{\link{getwts}}
}

