% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wtssrch_fast.R
\name{wtssrch_fast}
\alias{wtssrch_fast}
\alias{wtssrch_fast.default}
\title{Quickly find the optimal half-window width combination}
\usage{
wtssrch_fast(dat_in, ...)

\method{wtssrch_fast}{default}(dat_in, lower = c(0.2, 2, 0.2), upper = c(2,
  15, 2), ...)
}
\arguments{
\item{dat_in}{input data object to use with weighted regression}

\item{...}{arguments passed to or from other methods}

\item{lower}{vector of minimum half-window widths to evaluate}

\item{upper}{vector of maximum half-window widths to evaluate}
}
\value{
Some stuff
}
\description{
Quickly find the optimal half-window width combination to use for weighted regression
}
\details{
This is very similar to \code{\link{wtssrch}} but this function uses \code{\link[stats]{optim}} to minimize the error returned by \code{\link{wrtdscv}} for a given window combination.  The search algorithm uses the limited-memory modification of the BFGS quasi-Newton method to impose upper and lower limits on the optimization search. These limits can be changed using the \code{lower} and \code{upper} arguments.
}
\examples{
\dontrun{
# setup parallel backend
library(doParallel)
ncores <- detectCores() - 1
registerDoParallel(cores = ncores)

# run search function using default search grid - takes a while
res <- wtssrch_fast(tidobjmean)
}
}
\seealso{
\code{\link{wrtdscv}}, \code{\link{wtssrch}}
}

