% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chlinterp.R
\name{chlinterp}
\alias{chlinterp}
\alias{chlinterp.default}
\title{Interpolate from a fit grid}
\usage{
chlinterp(date_in, sal_in, ...)

\method{chlinterp}{default}(date_in, sal_in, fit_grd, sal_grd, ...)
}
\arguments{
\item{date_in}{obseved data to interpolate for chlorophyll}

\item{sal_in}{observed salinity value to interpolate for chlorophyll}

\item{...}{arguments passed to additional methods}

\item{fit_grd}{interpolation grid created from \code{\link{wrtds}}}

\item{sal_grd}{original vector of salinity values used to create interpolation grid}
}
\value{
Predicted chlorophyll values from the interpolation
}
\description{
Interpolate chlorophyll in two dimensions from the fit grid, where the first dimension is salinity and the second is date.
}
\details{
This function is used in \code{\link{chlpred}} and \code{\link{chlnorm}}.
}
\examples{
##

# load data
data(tidobj)

# interpolation grid from the model
fit_grd <- attr(tidobj, 'fits')
fit_grd <- fit_grd[['fit0.5']]

# salinity values used for interpolation grid
sal_grd <- attr(tidobj, 'sal_grd')

# get observed salinity value to inteprolate chlorophyll
date_in <- tidobj$date[1]
sal_in <- tidobj$sal[1]

# interpolate chlorophyll from an observed date and salinity value
chlinterp(date_in, sal_in, fit_grd, sal_grd)
}

