% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chlnorm.R
\name{chlnorm}
\alias{chlnorm}
\alias{chlnorm.tidal}
\title{Get salinity normalized WRTDS predictions from interpolation grids}
\usage{
chlnorm(tidal_in, ...)

\method{chlnorm}{tidal}(tidal_in, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to other methods}
}
\value{
Appends columns to the data.frame of normalized chlorophyll values for each quantile.  Columns are named starting with the prefix `norm', e.g., `norm0.5' are the normalized values for the fit through the median.
}
\description{
Get normalized model predictions from WRTDS to remove the effect of salinity on chlorophyll.  Predicted values in the interpolation grids are averaged across dates.
}
\details{
This function is used after \code{wrtds} to normalize predicted values of chlorophyll from the interpolation grid of each estimated quantile.  The normalized values are based on the average of all predicted chlorophyll estimates across the range of salinity values that have occurred on the same date throughout each year.  For example, normalized values for July 2000 are the mean predicted chlorophyll at that date using the observed salinity values that occur in July of all years.  The normalized values allow an interpretation of chlorophyll trends that are independent of changes in salinity or freshwater inputs.
}
\examples{
##

# load a tidal object
data(tidobj)

# get fitted values for each quantile
res <- chlnorm(tidobj)

##
\dontrun{
# fails if fits attribute not available
data(chldat)
chldat <- tidal(chldat)
chlnorm(chldat)
}
}

