% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chlpred.R
\name{chlpred}
\alias{chlpred}
\alias{chlpred.tidal}
\title{Get WRTDS predictions from interpolation grids}
\usage{
chlpred(tidal_in, ...)

\method{chlpred}{tidal}(tidal_in, ...)
}
\arguments{
\item{tidal_in}{input tidal object}
}
\value{
Appends columns to the data.frame of predicted chlorophyll values for each quantile.  Columns are named starting with the prefix `fit', e.g., `fit0.5' are the predicted values for the fit through the median.
}
\description{
Get model predictions from WRTDS using linear interpolation of values in grids
}
\details{
This function is used after \code{wrtds} to estimate predicted values of chlorophyll from the interpolation grid of each estimated quantile.  The estimated values are based on a linear interpolation of the two predicted chlorophyll values in the interpolation grid that are bounded by the upper and lower bounds of salinity for the observed salinity at each observation.
}
\examples{
##

# load a tidal object
data(tidobj)

# get fitted values for each quantile
res <- chlpred(tidobj)

##
\dontrun{
# fails if fits attribute not available
data(chldat)
chldat <- tidal(chldat)
chlpred(chldat)
}
}

