% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fill_grd.R
\name{fill_grd}
\alias{fill_grd}
\title{Fill an interpolation grid using all year, months}
\usage{
fill_grd(grd_in, dat_in)
}
\arguments{
\item{grd_in}{interpolation grid to fill, either a single mean grid or an individual quantile grid created within \code{\link{wrtds}}}

\item{dat_in}{\code{\link{tidal}} or \code{\link{tidalmean}} object}
}
\description{
Fill an interpolation grid using all year, months within the date range of the original dataset.  This is necessary for predicting data that may not be present in the observed data, as in a user-supplied dataset to the \code{dat_pred} argument in \code{\link{chlpred}}.  Function is used internally in \code{\link{wrtds}}.
}

