% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fill_grd.R
\name{fill_grd}
\alias{fill_grd}
\title{Fill missing values in the interpolation grid}
\usage{
fill_grd(grd_in, dat_in)
}
\arguments{
\item{grd_in}{interpolation grid to fill, either a single mean grid or an individual quantile grid created within \code{\link{wrtds}}}

\item{dat_in}{\code{\link{tidal}} or \code{\link{tidalmean}} object}
}
\description{
Fill missing values in the interpolation grid, only relevant for quantile models that describe the tails of the distribution.  Date, year, month, and day columns are also added using dates from \code{dat_in}.  Function is used in \code{\link{wrtds}}.
}

