% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_grd.R
\name{fill_grd}
\alias{fill_grd}
\title{Add additional date columns to the interpolation grids.}
\usage{
fill_grd(grd_in, dat_in)
}
\arguments{
\item{grd_in}{interpolation grid to fill, either a single mean grid or an individual quantile grid created within \code{\link{wrtds}}}

\item{dat_in}{\code{\link{tidal}} or \code{\link{tidalmean}} object}
}
\description{
Add date, year, month, and day columns to the interpolation grids using dates from \code{dat_in}.  Function is used in \code{\link{wrtds}}.
}

