% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitplot.R
\name{fitplot}
\alias{fitplot}
\alias{fitplot.tidal}
\title{Plot a tidal object}
\usage{
fitplot(tidal_in, ...)

\method{fitplot}{tidal}(tidal_in, tau = NULL, predicted = TRUE,
  annuals = FALSE, logspace = FALSE, pretty = TRUE, pt_sz = 2, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to \code{\link[ggplot2]{geom_line}}}

\item{tau}{numeric vector of quantiles to plot, defaults to all in object if not supplied}

\item{predicted}{logical indicating if standard predicted values are plotted, default \code{TRUE}, otherwise normalized predictions are plotted}

\item{annuals}{logical indicating if plots are annual aggregations of results}

\item{logspace}{logical indicating if plots are in log space}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}

\item{pt_sz}{numeric value indicating size of observed chlorophyll points}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot a tidal object to view chlorophyll observations, predictions, and normalized results.
}
\examples{
## load a fitted tidal object
data(tidfit)

# plot using defaults
fitplot(tidfit)

# get the same plot but use default ggplot settings
fitplot(tidfit, pretty = FALSE)

# plot in log space
fitplot(tidfit, logspace = TRUE)

# plot specific quantiles
fitplot(tidfit, tau = c(0.1, 0.9))

# plot the normalized predictions
fitplot(tidfit, predicted = FALSE)

# plot as annual aggregations
fitplot(tidfit, annuals = TRUE)

# format the x-axis is using annual aggregations
library(ggplot2)

fitplot(tidfit, annual = TRUE) +
 scale_x_continuous(breaks = seq(2000, 2012, by = 4))

# modify the plot as needed using ggplot scales, etc.

fitplot(tidfit, pretty = FALSE, linetype = 'dashed') +
 theme_classic() +
 scale_y_continuous(
   'Chlorophyll',
   limits = c(0, 50)
   ) +
 scale_colour_manual(
   'Predictions',
   labels = c('lo', 'md', 'hi'),
   values = c('red', 'green', 'blue'),
   guide = guide_legend(reverse = TRUE)
   )
}
\seealso{
\code{\link{prdnrmplot}}, \code{\link{sliceplot}}
}

