% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridplot.R
\name{gridplot}
\alias{gridplot}
\alias{gridplot.tidal}
\alias{gridplot.tidalmean}
\title{Plot chlorophyll response to salinity as a gridded surface for all months}
\usage{
gridplot(dat_in, ...)

\method{gridplot}{tidal}(dat_in, month = c(1:12), tau = NULL,
  years = NULL, col_vec = NULL, logspace = FALSE, allsal = FALSE,
  interp = TRUE, sal_fac = 3, yr_fac = sal_fac, ncol = NULL,
  grids = FALSE, pretty = TRUE, ...)

\method{gridplot}{tidalmean}(dat_in, month = c(1:12), years = NULL,
  col_vec = NULL, logspace = FALSE, allsal = FALSE, interp = TRUE,
  sal_fac = 3, yr_fac = sal_fac, ncol = NULL, grids = FALSE,
  pretty = TRUE, ...)
}
\arguments{
\item{dat_in}{input tidal or tidalmean object}

\item{...}{arguments passed to other methods}

\item{month}{numeric indicating months to plot or chr string 'all' to indicate all months with no plot facets}

\item{tau}{numeric vector of quantile to plot.  The function will plot the 'middle' quantile if none is specified, e.g., if 0.2, 0.3, and 0.4 are present in the fitted model object then 0.3 will be plotted.}

\item{years}{numeric vector of years to plot, defaults to all}

\item{col_vec}{chr string of plot colors to use, passed to \code{\link{gradcols}} and \code{\link[ggplot2]{scale_fill_gradientn}} for grid shading.  Any color palette from RColorBrewer can be used as a named input. Palettes from grDevices must be supplied as the returned string of colors for each palette.}

\item{logspace}{logical indicating if plots are in log space}

\item{allsal}{logical indicating if the salinity values for plotting are limited to the fifth and ninety-fifth percentile of observed salinity values for the month of interest}

\item{interp}{logical indicating if chlorophyll between integer years and salinity is linearly interpolated to create a smoother grid}

\item{sal_fac}{numeric value indicating the factor for smoothing chlorophyll across salinity values. Increasing the value creates more smoothing and setting the value to 1 removes all smoothing.}

\item{yr_fac}{numeric value indicating the factor for smoothing chlorophyll across integer years. Increasing the value creates more smoothing and setting the value to 1 removes all smoothing.}

\item{ncol}{numeric argument passed to \code{\link[ggplot2]{facet_wrap}} indicating number of facet columns}

\item{grids}{logical indicating if grid lines are present}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot the relationship between chlorophyll and salinity across the time series using a gridded surface for all months.  Chlorophyll is shaded by relative values across all dates for comparison.
}
\details{
These plots can be used to examine how the relationship between chlorophyll and salinity varies throughout the time series for multiple months.  The plot is similar to that returned by \code{\link{dynaplot}} except the chlorophyll changes are shown on a gridded surface of salinity versus time.  Multiple months can also be viewed.  Color shading is in proportion to the chlorophyll value and is relative across the plotted months.  The interpolation grid that is stored as an attribute in a fitted tidal object is used to create the plot.  By default, the plots are constrained to the fifth and ninety-fifth percentile of observed salinity values during each month to limit the predictions within the data domain. This behavior can be suppressed by changing the \code{allsal} argument, although the predicted chlorophyll values that are outside of the salinity range for the plotted month are typically unrealistic.
}
\examples{
\dontrun{
## load a fitted tidal object
data(tidfit)

## defaults to the fiftieth quantile
gridplot(tidfit)

## no facets, all months
gridplot(tidfit, months = 'all')

## change the defaults
gridplot(tidfit, tau = c(0.1), month = c(3, 6, 9, 12),
 col_vec = c('red', 'blue', 'green'), sal_fac = 1)

## plot a tidalmean object
data(tidfitmean)

gridplot(tidfitmean)

}
}
\seealso{
\code{\link{dynaplot}}, \code{\link{fitplot}}, \code{\link{gridplot}}, \code{\link{prdnrmplot}}
}

