% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/obsplot.R
\name{obsplot}
\alias{obsplot}
\alias{obsplot.tidal}
\title{Plot observed chlorophyll and salinity data}
\usage{
obsplot(tidal_in, ...)

\method{obsplot}{tidal}(tidal_in, lines = TRUE, logspace = FALSE,
  dt_rng = NULL, pretty = TRUE, col = "black", lwd = 1, size = 2,
  alpha = 1, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to \code{\link[ggplot2]{geom_line}}}

\item{lines}{logical indicating if a line plot is used, otherwise points}

\item{logspace}{logical indicating if plots are in log space}

\item{dt_rng}{Optional chr string indicating the date range of the plot. Must be two values in the format 'YYYY-mm-dd' which is passed to \code{\link{as.Date}}.}

\item{pretty}{logical indicating if my subjective idea of plot aesthetics is applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used
colvec = 'black',}

\item{col}{chr string of plot color to use}

\item{lwd}{numeric value indicating width of lines}

\item{size}{numeric value indicating size of points}

\item{alpha}{numeric value indicating transparency of points or lines}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot observed chlorophyll and salinity time series from a tidal object
}
\examples{
## load a fitted tidal object
data(tidfit)

# plot using defaults
obsplot(tidfit)

# changing default
obsplot(tidfit, alpha = 0.5, size = 4, col = 'blue', lines = FALSE)
}
\seealso{
\code{\link{fitplot}}
}

