% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resinterp.R
\name{resinterp}
\alias{resinterp}
\alias{resinterp.default}
\title{Interpolate from a fit grid}
\usage{
resinterp(date_in, sal_in, ...)

\method{resinterp}{default}(date_in, sal_in, fit_grd, sal_grd, ...)
}
\arguments{
\item{date_in}{obseved data to interpolate for the response variable}

\item{sal_in}{observed salinity value to interpolate for the response variable}

\item{...}{arguments passed to additional methods}

\item{fit_grd}{interpolation grid created from \code{\link{wrtds}}}

\item{sal_grd}{original vector of salinity values used to create interpolation grid}
}
\value{
Predicted values of the response variable from the interpolation
}
\description{
Interpolate the response variable in two dimensions from the fit grid, where the first dimension is salinity and the second is date.
}
\details{
This function is used in \code{\link{respred}} and \code{\link{resnorm}}.
}
\examples{
##

# load data
data(tidobj)

# interpolation grid from the model
fit_grd <- attr(tidobj, 'fits')
fit_grd <- fit_grd[['fit0.5']]

# salinity values used for interpolation grid
sal_grd <- attr(tidobj, 'sal_grd')

# get observed salinity value to interpolate the response
date_in <- tidobj$date[1]
sal_in <- tidobj$sal[1]

# interpolate the response from an observed date and salinity value
resinterp(date_in, sal_in, fit_grd, sal_grd)
}

