% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/salfind.R
\name{salfind}
\alias{salfind}
\alias{salfind.tidal}
\title{Get all salinity values across years for a common date}
\usage{
salfind(tidal_in, ...)

\method{salfind}{tidal}(tidal_in, obs, ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{...}{arguments passed to additional methods}

\item{obs}{index of row in tidal object to identify salff values}
}
\value{
Numeric vector of all observed salff values for a single month across all years
}
\description{
Get all salinity values across years for a common date for normalizing chlorophyll predictions
}
\details{
This function is primarily used within \code{\link{chlnorm}}.  It will return all observed salinity (as fraction of freshwater) values for a given month across years.  For example, the function will return all observed salff values for January across all years in the dataset if the input row is for January 2000.  An identical vector would result for the same month in other years.
}
\examples{
##

# load data
data(tidobj)

# find common salff values across years for the first observation
res <- salfind(tidobj, 1)
}

