% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/salfind.R
\name{salfind}
\alias{salfind}
\alias{salfind.tidal}
\title{Get all salinity values across years for a common date}
\usage{
salfind(dat_in, ...)

\method{salfind}{tidal}(dat_in, obs, ...)
}
\arguments{
\item{dat_in}{input tidal object}

\item{...}{arguments passed to additional methods}

\item{obs}{index of row in tidal object to identify salinity values}
}
\value{
Numeric vector of all observed salinity values for a single month across all years
}
\description{
Get all salinity values across years for a common date for normalizing chlorophyll predictions
}
\details{
This function is primarily used within \code{\link{chlnorm}}.  It will return all observed salinity values for a given date for matching months across all years.  For example, the function will return all observed salinity values for January across all years in the dataset if the input row is for January 2000.  An identical vector would result for the same month in other years.
}
\examples{
##

# load data
data(tidobj)

# find common salinity values across years for the first observation
res <- salfind(tidobj, 1)
}

