% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidal.R
\name{tidal}
\alias{tidal}
\title{Create a tidal class object}
\usage{
tidal(dat_in, ind = c(1, 2, 3), chllog = TRUE)
}
\arguments{
\item{dat_in}{Input data frame for a water quality time series with three columns for date (Y-m-d format), chlorophyll concentation (ug/L), and fraction of freshwater}

\item{chllog}{logical indicating if input chlorophyll is already in log-space, default \code{TRUE}}

\item{three}{element numeric vector indicating column positions of date, chlorophyll, and salinity of input data frame}
}
\value{
A \code{data.frame} with columns ordered as date, chlorophyll, salinity, and decimal time
}
\description{
Prepare water quality data for weighted regression by creating a tidal class object
}
\details{
This function is a simple wrapper to \code{\link[base]{structure}} that is used to create a tidal object for use with weighted regression in tidal waters. Input data should be a three-column \code{\link[base]{data.frame}} with date, chlorophyll, and salinity data.  Chlorophyll data are assumed to be log-transformed, otherwise use \code{chllog = FALSE}.  Salinity data should be represented as fraction of freshwater.
}
\examples{
## raw data

data(chldat)

## format
chldat <- tidal(chldat)
}

