% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrtds.R
\name{wrtds}
\alias{wrtds}
\alias{wrtds.data.frame}
\alias{wrtds.tidal}
\title{Get WRTDS prediction grid}
\usage{
wrtds(tidal_in, sal_div = 10, tau = 0.5, trace = TRUE, ...)

\method{wrtds}{data.frame}(tidal_in, sal_div = 10, tau = 0.5,
  trace = TRUE, ...)

\method{wrtds}{tidal}(tidal_in, sal_div = 10, tau = 0.5, trace = TRUE,
  ...)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{sal_div}{numeric indicating number of divisions across the range of salinity to create the interpolation grid}

\item{tau}{numeric vector indicating conitional quantiles to fit in the weighted regression, can be many}

\item{trace}{logical indicating if progress is shown in the console}

\item{...}{arguments passed to or from other methods}
}
\value{
Appends a regression grid as an attribute to a tidal object
}
\description{
Get WRTDS prediction grid for chlorophyll observations in a tidal object
}
\examples{
\dontrun{
## load data
data(chldat)

## get wrtds grid, as data frame input
res <- wrtds(chldat)

## as tidal object
tidal_in <- tidal(chldat)
res <- wrtds(tidal_in)

## multiple quantiles
res <- wrtds(chldat, tau = c(0.1, 0.5, 0.9))
}
}

