% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrtds.R
\name{wrtds}
\alias{wrtds}
\alias{wrtds.data.frame}
\alias{wrtds.tidal}
\title{Get WRTDS prediction grid}
\usage{
wrtds(tidal_in, sal_div = 20, ...)

\method{wrtds}{data.frame}(tidal_in, sal_div = 20, ...)

\method{wrtds}{tidal}(tidal_in, sal_div = 20)
}
\arguments{
\item{tidal_in}{input tidal object}

\item{sal_div}{numeric indicating number of divisions across the range of salinity as fraction of freshwater to create interpolation grid}

\item{...}{arguments passed to other methods}
}
\value{
Appends a regression grid as an attribute to a tidal object
}
\description{
Get WRTDS prediction grid for chlorophyll observations in a tidal object
}
\examples{
## load data
data(chldat)

## get wrtds grid
res <- wrtds(chldat)
}

