% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrtds.R
\name{wrtds}
\alias{wrtds}
\alias{wrtds.tidal}
\alias{wrtds.tidalmean}
\title{Get WRTDS prediction grid}
\usage{
wrtds(dat_in, ...)

\method{wrtds}{tidal}(dat_in, sal_div = 10, tau = 0.5, trace = TRUE, ...)

\method{wrtds}{tidalmean}(dat_in, sal_div = 10, trace = TRUE, ...)
}
\arguments{
\item{dat_in}{input tidal or tidalmean object}

\item{...}{arguments passed to or from other methods}

\item{sal_div}{numeric indicating number of divisions across the range of salinity to create the interpolation grid}

\item{tau}{numeric vector indicating conitional quantiles to fit in the weighted regression, can be many}

\item{trace}{logical indicating if progress is shown in the console}
}
\value{
Appends interpolation grid attributes to the input object.  For a tidal object, this will include multiple grids for each quantile.  For tidalmean objects, only one grid is appended to the `fits' attribute, in addition a back-transformed grid for the `bt_fits' attribute.
}
\description{
Get WRTDS prediction grid for chlorophyll observations in a tidal or tidalmean object
}
\examples{
\dontrun{
## load data
data(chldat)

## as tidal object
dat_in <- tidal(chldat)
res <- wrtds(dat_in)

## as tidalmean object
dat_in <- tidalmean(chldat)
res <- wrtds(dat_in)

## multiple quantiles
res <- wrtds(chldat, tau = c(0.1, 0.5, 0.9))
}
}

