% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrtdscv.R
\name{wrtdscv}
\alias{wrtdscv}
\alias{wrtdscv.default}
\title{Use k-fold cross-validation to evaluate WRTDS model fit}
\usage{
wrtdscv(dat_in, ...)

\method{wrtdscv}{default}(dat_in, wins_in, k = 10, trace = TRUE,
  seed_val = NULL, ...)
}
\arguments{
\item{dat_in}{input model object}

\item{...}{arguments passed to or from other methods (e.g., \code{\link{getwts}})}

\item{wins_in}{list of input half-window widths of the order months, years, and salinity, passed to \code{\link{getwts}}}

\item{k}{number of folds to evaluate}

\item{trace}{logical indicating if progress is printed in the console}
}
\value{
some more crap
}
\description{
Use k-fold cross-validation to evaluate WRTDS model fit
}
\details{
Some details
}
\examples{
\dontrun{

# half-window widths to evaluate
# months, years, and salinity
wins <- list(0.5, 10, 0.5)

#get ocv score for k = 10
wrtdscv(tidobjmean, wins_in = wins)

}
}
\seealso{
\code{\link{getwts}}, \code{\link{wtsplot}}
}

