% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wtssrch_fast.R
\name{wtssrch_fast}
\alias{wtssrch_fast}
\alias{wtssrch_fast.default}
\title{Quickly find the optimal half-window width combination}
\usage{
wtssrch_fast(dat_in, ...)

\method{wtssrch_fast}{default}(dat_in, wins_in = NULL, trace = TRUE, ...)
}
\arguments{
\item{dat_in}{input data object to use with weighted regression}

\item{...}{arguments passed to or from other methods}

\item{wins_in}{starting list of window weights for initializing the search algorithm}

\item{trace}{logical indicating if progress is saved to a text file in the working directory}
}
\value{
Some stuff
}
\description{
Quickly find the optimal half-window width combination to use for weighted regression
}
\details{
This is very similar to \code{\link{wtssrch}} but this function provides a quicker alternative by using the \code{\link[stats]{optim}} function to minimize the error returned by \code{\link{wrtdscv}} for a given window combination.
}
\examples{
\dontrun{
# setup parallel backend
library(doParallel)
ncores <- detectCores() - 1
registerDoParallel(cores = ncores)

# run search function using default search grid - takes a while
res <- wtssrch_fast(tidobjmean)
}
}
\seealso{
\code{\link{wrtdscv}}, \code{\link{wtssrch}}
}

